requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
idna==2.9
jira==2.0.0
multidict==4.7.5
oauthlib==3.1.0
pbr==5.4.4
pycparser==2.19
PyJWT==1.7.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.14.0
slackbot==1.0.0
slackclient==2.5.0
slacker==0.14.0
typing_extensions==4.1.1
urllib3==1.25.8
websocket-client==0.44.0
yarl==1.4.2
